chore: Extract traits from Share/Unshare collection forms
