+++ src/forms/collections/ShareCollection.php
@@ -3,10 +3,9 @@
 namespace App\forms\collections;
 
 use App\forms\BaseForm;
+use App\forms\traits;
 use App\models;
-use App\utils;
 use Minz\Form;
-use Minz\Request;
 use Minz\Translatable;
 use Minz\Validable;
 
@@ -18,7 +17,7 @@ use Minz\Validable;
  */
 class ShareCollection extends BaseForm
 {
-    use utils\Memoizer;
+    use traits\ShareRecipient;
 
     #[Form\Field]
     #[Validable\Presence(
@@ -29,16 +28,6 @@ class ShareCollection extends BaseForm
         message: new Translatable('The type is invalid.'),
     )]
     public string $type = 'read';
-
-    #[Form\Field(transform: 'trim')]
-    public string $user_id = '';
-
-    public function user(): models\User
-    {
-        return $this->memoize('user', function (): models\User {
-            return models\User::require($this->user_id);
-        });
-    }
 
     /**
      * @return ShareType
@@ -52,49 +41,25 @@ class ShareCollection extends BaseForm
         return $this->type;
     }
 
-    #[Form\OnHandleRequest]
-    public function extractUserIdFromProfileUrl(Request $request): void
-    {
-        list($origin_type, $origin_id) = utils\OriginHelper::extractFromPath($this->user_id);
-
-        if ($origin_type === 'user' && $origin_id) {
-            $this->user_id = $origin_id;
-        }
-    }
-
-    #[Validable\Check]
-    public function checkUserIdIsValid(): void
+    protected function isRecipientOwner(models\User $user): bool
     {
         $collection = $this->optionAs('collection', models\Collection::class);
+        return $collection->user_id === $user->id;
+    }
 
-        if (!models\User::exists($this->user_id)) {
-            $this->addError(
-                'user_id',
-                'user_id.unknown',
-                _('This user doesn’t exist.'),
-            );
-            return;
-        }
+    protected function isAlreadySharedWith(models\User $user): bool
+    {
+        $collection = $this->optionAs('collection', models\Collection::class);
+        return $collection->sharedWith($user);
+    }
 
+    protected function recipientIsOwnerError(): string
+    {
+        return _('You can’t share access with the owner of the collection.');
+    }
 
-        if ($collection->user_id === $this->user_id) {
-            $this->addError(
-                'user_id',
-                'user_id.same_as_owner',
-                _('You can’t share access with the owner of the collection.'),
-            );
-            return;
-        }
-
-        $user = $this->user();
-
-        if ($collection->sharedWith($user)) {
-            $this->addError(
-                'user_id',
-                'user_id.already_shared',
-                _('The collection is already shared with this user.'),
-            );
-            return;
-        }
+    protected function alreadySharedError(): string
+    {
+        return _('The collection is already shared with this user.');
     }
 }
