+++ src/forms/collections/UnshareCollection.php
@@ -3,9 +3,7 @@
 namespace App\forms\collections;
 
 use App\forms\BaseForm;
-use App\models;
-use Minz\Form;
-use Minz\Validable;
+use App\forms\traits;
 
 /**
  * @author  Marien Fressinaud <dev@marienfressinaud.fr>
@@ -13,24 +11,5 @@ use Minz\Validable;
  */
 class UnshareCollection extends BaseForm
 {
-    #[Form\Field(transform: 'trim')]
-    public string $user_id = '';
-
-    public function user(): models\User
-    {
-        return models\User::require($this->user_id);
-    }
-
-    #[Validable\Check]
-    public function checkUserIdIsValid(): void
-    {
-        if (!models\User::exists($this->user_id)) {
-            $this->addError(
-                '@base',
-                'user_id.unknown',
-                _('This user doesn’t exist.'),
-            );
-            return;
-        }
-    }
+    use traits\UnshareRecipient;
 }
